# api/openapi/swagger.yaml
components:
  schemas:
    errs.Error:
      properties:
        code:
          $ref: '#/components/schemas/errs.ErrorCode'
        message:
          type: string
        params:
          items:
            $ref: '#/components/schemas/errs.Param'
          type: array
          uniqueItems: false
      type: object
    errs.ErrorCode:
      type: integer
      x-enum-varnames:
      - ErrorCodeOK
      - ErrorCodeCanceled
      - ErrorCodeUnknown
      - ErrorCodeInvalidArgument
      - ErrorCodeDeadlineExceeded
      - ErrorCodeNotFound
      - ErrorCodeAlreadyExists
      - ErrorCodePermissionDenied
      - ErrorCodeResourceExhausted
      - ErrorCodeFailedPrecondition
      - ErrorCodeAborted
      - ErrorCodeOutOfRange
      - ErrorCodeUnimplemented
      - ErrorCodeInternal
      - ErrorCodeUnavailable
      - ErrorCodeDataLoss
      - ErrorCodeUnauthenticated
      - ErrorCodeClosedRequest
    errs.Param:
      properties:
        key:
          type: string
        value:
          type: string
      type: object
    handlers.ArticleCreateDTO:
      properties:
        body:
          type: string
        is_published:
          type: boolean
        subtitle:
          type: string
        title:
          type: string
      type: object
    handlers.ArticleDTO:
      properties:
        body:
          type: string
        created_at:
          type: string
        id:
          type: string
        is_published:
          type: boolean
        subtitle:
          type: string
        title:
          type: string
        updated_at:
          type: string
      type: object
    handlers.ArticleListDTO:
      properties:
        count:
          type: integer
        items:
          items:
            $ref: '#/components/schemas/handlers.ArticleDTO'
          type: array
          uniqueItems: false
      type: object
    handlers.ArticleUpdateDTO:
      properties:
        body:
          type: string
        id:
          type: string
        is_published:
          type: boolean
        subtitle:
          type: string
        title:
          type: string
      type: object
...
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  title: example
  version: 0.0.0
openapi: 3.1.0
paths:
  /api/v1/articles/articles/:
    get:
      parameters:
      - in: query
        name: page_number
        schema:
          type: integer
      - in: query
        name: order_by
        schema:
          items:
            type: string
          type: array
          uniqueItems: false
      - in: query
        name: search
        schema:
          type: string
      - in: query
        name: page_size
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ArticleListDTO'
          description: Filtered list of articles
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: List of articles
      tags:
      - article
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.ArticleCreateDTO'
        description: Create article request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ArticleDTO'
          description: Created article
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Create article
      tags:
      - article
  /api/v1/articles/articles/{id}:
    delete:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          description: No content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Delete article by id
      tags:
      - article
    get:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ArticleDTO'
          description: Requested article
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Get article by id
      tags:
      - article
    patch:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.ArticleUpdateDTO'
        description: Update article request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ArticleDTO'
          description: Updated article
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Update article
      tags:
      - article
...
servers:
- url: http://127.0.0.1:8000/